requirements.txt
aiofiles==24.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
click==8.1.8
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0

passlib~=1.7.4
python-jose~=3.3.0